--- sample_config.ini
@@ -1,8 +1,0 @@
-[iron_mq]
-token = ENTER_YOUR_TOKEN_HERE
-project_id = ENTER_YOUR_PROJECT_ID_HERE
-; Optional fields:
-; protocol = http
-; api_version = 1
-; host = mq-aws-us-east-1.iron.io
-; port = 80
