+++ src/IronMQ.php
@@ -1,4 +1,9 @@
 <?php
+
+namespace IronMQ;
+
+use IronCore\IronCore;
+
 /**
  * PHP client for IronMQ
  * IronMQ is a scalable, reliable, high performance message queue in the cloud.
@@ -6,141 +11,14 @@
  * @link https://github.com/iron-io/iron_mq_php
  * @link http://www.iron.io/products/mq
  * @link http://dev.iron.io/
- * @version 3.0.2
+ * @version 4.0.0
  * @package IronMQPHP
  * @copyright Feel free to copy, steal, take credit for, or whatever you feel like doing with this code. ;)
  */
-
-
-if (!class_exists('IronCore')) {
-    if (!class_exists('Composer\Autoload\ClassLoader')) {
-        echo "Please include IronCore class first\n";
-    }
-    return;
-}
-
-class IronMQ_Exception extends Exception
-{
-
-}
-
-
-class IronMQ_Message
-{
-    private $body;
-    private $timeout;
-    private $delay;
-    private $expires_in;
-
-    const MAX_EXPIRES_IN = 2592000;
-
-    /**
-     * Create a new message.
-     *
-     * @param string $message
-     *        A message body
-     * @param array $properties
-     *        An array of message properties
-     * Fields in $properties array:
-     * - timeout: Timeout, in seconds. After timeout, item will be placed back on queue. Defaults to 60.
-     * - delay: The item will not be available on the queue until this many seconds have passed. Defaults to 0.
-     * - expires_in: How long, in seconds, to keep the item on the queue before it is deleted.
-     *               Defaults to 604800 (7 days). Maximum is 2592000 (30 days).
-     */
-    public function __construct($message, $properties = array())
-    {
-        $this->setBody($message);
-
-        if (array_key_exists("timeout", $properties)) {
-            $this->setTimeout($properties['timeout']);
-        }
-        if (array_key_exists("delay", $properties)) {
-            $this->setDelay($properties['delay']);
-        }
-        if (array_key_exists("expires_in", $properties)) {
-            $this->setExpiresIn($properties['expires_in']);
-        }
-    }
-
-    public function getBody()
-    {
-        return $this->body;
-    }
-
-    public function setBody($body)
-    {
-        if (empty($body)) {
-            throw new InvalidArgumentException("Please specify a body");
-        } else {
-            $this->body = (string) $body;
-        }
-    }
-
-    public function getTimeout()
-    {
-        # 0 is considered empty, but we want people to be able to set a timeout of 0
-        if (!empty($this->timeout) || $this->timeout === 0) {
-            return $this->timeout;
-        } else {
-            return null;
-        }
-    }
-
-    public function setTimeout($timeout)
-    {
-        $this->timeout = $timeout;
-    }
-
-    public function getDelay()
-    {
-        # 0 is considered empty, but we want people to be able to set a delay of 0
-        if (!empty($this->delay) || $this->delay == 0) {
-            return $this->delay;
-        } else {
-            return null;
-        }
-    }
-
-    public function setDelay($delay)
-    {
-        $this->delay = $delay;
-    }
-
-    public function getExpiresIn()
-    {
-        return $this->expires_in;
-    }
-
-    public function setExpiresIn($expires_in)
-    {
-        if ($expires_in > self::MAX_EXPIRES_IN) {
-            throw new InvalidArgumentException("Expires In can't be greater than ".self::MAX_EXPIRES_IN.".");
-        } else {
-            $this->expires_in = $expires_in;
-        }
-    }
-
-    public function asArray()
-    {
-        $array = array();
-        $array['body'] = $this->getBody();
-        if ($this->getTimeout() != null) {
-            $array['timeout'] = $this->getTimeout();
-        }
-        if ($this->getDelay() != null) {
-            $array['delay'] = $this->getDelay();
-        }
-        if ($this->getExpiresIn() != null) {
-            $array['expires_in'] = $this->getExpiresIn();
-        }
-        return $array;
-    }
-}
-
 class IronMQ extends IronCore
 {
 
-    protected $client_version = '3.0.2';
+    protected $client_version = '4.0.0';
     protected $client_name    = 'iron_mq_php';
     protected $product_name   = 'iron_mq';
     protected $default_values = array(
@@ -182,10 +60,12 @@ class IronMQ extends IronCore
      */
     public function setProjectId($project_id)
     {
-        if (!empty($project_id)) {
+        if (!empty($project_id))
+        {
             $this->project_id = $project_id;
         }
-        if (empty($this->project_id)) {
+        if (empty($this->project_id))
+        {
             throw new InvalidArgumentException("Please set project_id");
         }
     }
@@ -202,10 +82,12 @@ class IronMQ extends IronCore
     {
         $url = "projects/{$this->project_id}/queues";
         $params = array();
-        if (!is_null($previous)) {
+        if (!is_null($previous))
+        {
             $params['previous'] = $previous;
         }
-        if ($per_page !== self::LIST_QUEUES_PER_PAGE) {
+        if ($per_page !== self::LIST_QUEUES_PER_PAGE)
+        {
             $params['per_page'] = (int) $per_page;
         }
         $this->setJsonHeaders();
@@ -263,7 +145,7 @@ class IronMQ extends IronCore
      */
     public function postMessage($queue_name, $message, $properties = array())
     {
-        $msg = new IronMQ_Message($message, $properties);
+        $msg = new IronMQMessage($message, $properties);
         $req = array(
             "messages" => array($msg->asArray())
         );
@@ -298,8 +180,9 @@ class IronMQ extends IronCore
         $req = array(
             "messages" => array()
         );
-        foreach ($messages as $message) {
-            $msg = new IronMQ_Message($message, $properties);
+        foreach ($messages as $message)
+        {
+            $msg = new IronMQMessage($message, $properties);
             array_push($req['messages'], $msg->asArray());
         }
         $this->setJsonHeaders();
@@ -336,21 +219,27 @@ class IronMQ extends IronCore
         $queue = rawurlencode($queue_name);
         $url = "projects/{$this->project_id}/queues/$queue/reservations";
         $params = array();
-        if ($count !== 1) {
+        if ($count !== 1)
+        {
             $params['n'] = (int) $count;
         }
-        if ($timeout !== self::GET_MESSAGE_TIMEOUT) {
+        if ($timeout !== self::GET_MESSAGE_TIMEOUT)
+        {
             $params['timeout'] = (int) $timeout;
         }
-        if ($wait !== self::GET_MESSAGE_WAIT) {
+        if ($wait !== self::GET_MESSAGE_WAIT)
+        {
             $params['wait'] = (int) $wait;
         }
         $this->setJsonHeaders();
         $response = $this->apiCall(self::POST, $url, $params);
         $result = self::json_decode($response);
-        if (count($result->messages) < 1) {
+        if (count($result->messages) < 1)
+        {
             return null;
-        } else {
+        }
+        else
+        {
             return $result->messages;
         }
     }
@@ -378,9 +267,12 @@ class IronMQ extends IronCore
     public function reserveMessage($queue_name, $timeout = self::GET_MESSAGE_TIMEOUT, $wait = self::GET_MESSAGE_WAIT)
     {
         $messages = $this->reserveMessages($queue_name, 1, $timeout, $wait);
-        if ($messages) {
+        if ($messages)
+        {
             return $messages[0];
-        } else {
+        }
+        else
+        {
             return null;
         }
     }
@@ -416,9 +308,12 @@ class IronMQ extends IronCore
         $this->setJsonHeaders();
         $queue = rawurlencode($queue_name);
         $url = "projects/{$this->project_id}/queues/$queue/messages/{$message_id}";
-        if (is_null($reservation_id)) {
+        if (is_null($reservation_id))
+        {
             return $this->apiCall(self::DELETE, $url);
-        } else {
+        }
+        else
+        {
             return $this->apiCall(self::DELETE, $url, $req);
         }
     }
@@ -437,12 +332,18 @@ class IronMQ extends IronCore
         $req = array(
             "ids" => array()
         );
-        foreach ($messages as $message) {
-            if (is_string($message)) {
+        foreach ($messages as $message)
+        {
+            if (is_string($message))
+            {
                 array_push($req['ids'], $message);
-            } else if (is_object($message)) {
+            }
+            else if (is_object($message))
+            {
                 array_push($req['ids'], array('id' => $message->id, 'reservation_id' => $message->reservation_id));
-            } else if (is_array($message)) {
+            }
+            else if (is_array($message))
+            {
                 array_push($req['ids'], array('id' => $message['id'], 'reservation_id' => $message['reservation_id']));
             }
         }
@@ -463,9 +364,12 @@ class IronMQ extends IronCore
     public function peekMessage($queue_name)
     {
         $messages = $this->peekMessages($queue_name, 1);
-        if ($messages == null) {
+        if ($messages == null)
+        {
             return null;
-        } else {
+        }
+        else
+        {
             return $messages[0];
         }
     }
@@ -483,7 +387,8 @@ class IronMQ extends IronCore
         $queue = rawurlencode($queue_name);
         $url = "projects/{$this->project_id}/queues/$queue/messages";
         $params = array();
-        if ($count !== 1) {
+        if ($count !== 1)
+        {
             $params['n'] = (int) $count;
         }
         $this->setJsonHeaders();
@@ -506,7 +411,8 @@ class IronMQ extends IronCore
         $req = array(
             "reservation_id" => $reservation_id
         );
-        if ($timeout !== 0) {
+        if ($timeout !== 0)
+        {
             $req['timeout'] = (int) $timeout;
         }
         $this->setJsonHeaders();
@@ -532,7 +438,8 @@ class IronMQ extends IronCore
         $this->setJsonHeaders();
         $queue = rawurlencode($queue_name);
         $params = array('reservation_id' => $reservation_id);
-        if ($delay !== 0) {
+        if ($delay !== 0)
+        {
             $params['delay'] = (int) $delay;
         }
         $url = "projects/{$this->project_id}/queues/$queue/messages/{$message_id}/release";
@@ -808,12 +715,12 @@ class IronMQ extends IronCore
     {
         $this->setCommonHeaders();
         $token = isset($this->use_keystone) && $this->use_keystone ? $this->getToken(): $this->token;
-        $this->headers['Authorization'] ="OAuth {$token}";
+        $this->headers['Authorization'] = "OAuth {$token}";
     }
 
     private function setPostHeaders()
     {
         $this->setCommonHeaders();
-        $this->headers['Content-Type'] ='multipart/form-data';
+        $this->headers['Content-Type'] = 'multipart/form-data';
     }
 }
