+++ src/IronMQException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace IronMQ;
+
+/**
+ * IronMQException
+ * @package IronMQ
+ * @author Tino Ehrich (tino@bigpun.me)
+ */
+class IronMQException extends \Exception
+{
+}
